Pull #12102: removed custom cache for resource bundle cache for messages
