+++ .ci/checker-framework-suppressions/checker-framework-formatter-suppressions.xml
@@ -1,14 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <suppressedErrors>
   <checkerFrameworkError unstable="false">
-    <fileName>checkstyle/src/main/java/com/puppycrawl/tools/checkstyle/DefaultLogger.java</fileName>
-    <specifier>i18nformat.key.not.found</specifier>
-    <message>a key doesn&apos;t exist in the provided translation file</message>
-    <lineContent>final String pattern = resourceBundle.getString(key);</lineContent>
-  </checkerFrameworkError>
-
-  <checkerFrameworkError unstable="false">
-    <fileName>checkstyle/src/main/java/com/puppycrawl/tools/checkstyle/api/Violation.java</fileName>
+    <fileName>checkstyle/src/main/java/com/puppycrawl/tools/checkstyle/LocalizedMessage.java</fileName>
     <specifier>i18nformat.key.not.found</specifier>
     <message>a key doesn&apos;t exist in the provided translation file</message>
     <lineContent>final String pattern = resourceBundle.getString(key);</lineContent>
