+++ src/main/java/com/puppycrawl/tools/checkstyle/LocalizedMessage.java
@@ -1,0 +1,158 @@
+///////////////////////////////////////////////////////////////////////////////////////////////
+// checkstyle: Checks Java source code and other text files for adherence to a set of rules.
+// Copyright (C) 2001-2022 the original author or authors.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+///////////////////////////////////////////////////////////////////////////////////////////////
+
+package com.puppycrawl.tools.checkstyle;
+
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.Reader;
+import java.net.URL;
+import java.net.URLConnection;
+import java.nio.charset.StandardCharsets;
+import java.text.MessageFormat;
+import java.util.Arrays;
+import java.util.Locale;
+import java.util.MissingResourceException;
+import java.util.PropertyResourceBundle;
+import java.util.ResourceBundle;
+import java.util.ResourceBundle.Control;
+
+/**
+ * Represents a message that can be localised. The translations come from
+ * message.properties files. The underlying implementation uses
+ * java.text.MessageFormat.
+ */
+public class LocalizedMessage {
+
+    /** Name of the resource bundle to get messages from. **/
+    private final String bundle;
+
+    /**
+     * The locale to localise messages to.
+     **/
+    private final Locale locale;
+
+    /** Class of the source for this message. */
+    private final Class<?> sourceClass;
+
+    /**
+     * Key for the message format.
+     **/
+    private final String key;
+
+    /**
+     * Arguments for MessageFormat.
+     */
+    private final Object[] args;
+
+    /**
+     * Creates a new {@code LocalizedMessage} instance.
+     *
+     * @param bundle resource bundle name
+     * @param locale The locale to localise messages to.
+     * @param sourceClass the Class that is the source of the message
+     * @param key the key to locate the translation.
+     * @param args arguments for the translation.
+     */
+    public LocalizedMessage(String bundle, Locale locale, Class<?> sourceClass, String key,
+            Object... args) {
+        this.bundle = bundle;
+        this.locale = locale;
+        this.sourceClass = sourceClass;
+        this.key = key;
+        if (args == null) {
+            this.args = null;
+        }
+        else {
+            this.args = Arrays.copyOf(args, args.length);
+        }
+    }
+
+    /**
+     * Gets the translated message.
+     *
+     * @return the translated message.
+     */
+    public String getMessage() {
+        String result;
+        try {
+            // Important to use the default class loader, and not the one in
+            // the GlobalProperties object. This is because the class loader in
+            // the GlobalProperties is specified by the user for resolving
+            // custom classes.
+            final ResourceBundle resourceBundle = getBundle();
+            final String pattern = resourceBundle.getString(key);
+            final MessageFormat formatter = new MessageFormat(pattern, Locale.ROOT);
+            result = formatter.format(args);
+        }
+        catch (final MissingResourceException ignored) {
+            // If the Check author didn't provide i18n resource bundles
+            // and logs audit event messages directly, this will return
+            // the author's original message
+            final MessageFormat formatter = new MessageFormat(key, Locale.ROOT);
+            result = formatter.format(args);
+        }
+        return result;
+    }
+
+    /**
+     * Obtain the ResourceBundle. Uses the classloader
+     * of the class emitting this message, to be sure to get the correct
+     * bundle.
+     *
+     * @return a ResourceBundle.
+     */
+    private ResourceBundle getBundle() {
+        return ResourceBundle.getBundle(bundle, locale, sourceClass.getClassLoader(),
+                new Utf8Control());
+    }
+
+    /**
+     * <p>
+     * Custom ResourceBundle.Control implementation which allows explicitly read
+     * the properties files as UTF-8.
+     * </p>
+     */
+    public static class Utf8Control extends Control {
+
+        @Override
+        public ResourceBundle newBundle(String baseName, Locale locale, String format,
+                 ClassLoader loader, boolean reload) throws IOException {
+            // The below is a copy of the default implementation.
+            final String bundleName = toBundleName(baseName, locale);
+            final String resourceName = toResourceName(bundleName, "properties");
+            final URL url = loader.getResource(resourceName);
+            ResourceBundle resourceBundle = null;
+            if (url != null) {
+                final URLConnection connection = url.openConnection();
+                if (connection != null) {
+                    connection.setUseCaches(!reload);
+                    try (Reader streamReader = new InputStreamReader(connection.getInputStream(),
+                            StandardCharsets.UTF_8)) {
+                        // Only this line is changed to make it read property files as UTF-8.
+                        resourceBundle = new PropertyResourceBundle(streamReader);
+                    }
+                }
+            }
+            return resourceBundle;
+        }
+
+    }
+
+}
